Rewind Editor::doc_rainbow_highlights to containing syntax node
Here we rewind the doc_rainbow_highlights iterator to a stable point
for calculating the nesting level. The rainbow highlights iterator
can't start at the beginning of the viewport because then the nesting
level calculation would depend on the viewport which leads to color
changes as you scroll.

Where exactly do we start? We find the child of the syntax tree's
root node which contains the starting byte of the viewport. This
allows us to start not all the way back at the beginning of the
document but to skip ahead. For most tree-sitter grammars I've
seen and on most normal documents, this means that we'll be
running 1000 lines back in the document at most. Degenerate
cases exist when the root has few children and those children cover
a very large byte range.

There's room for future improvement where we can find the syntax
node highest in the tree which contains any of the nodes in the
patterns we search for in rainbows.scm, and we can start at the
beginning of that node. That approach would minimize how far back
in the document we need to start the iterator.