Add a "rainbow.include-children" property for HTML
This allows us to use rainbow nesting highlights for HTML
tags. The nesting node is the HTML grammar's `element` node.
`<`/`>` are direct descendants of `start_tag` and `end_tag` rather
than `element` though. `<`/`>` is not valid in this grammar except
in cases where we want to highlight it. So we can exclude the
`start_tag` and `end_tag` parents (which aren't scopes) and
instead have the `element` match on any children `<`/`>`.

Note that "rainbow.include-children" only applies to the scope
captured with `@rainbow.scope`: any children scopes will disable
the setting.